refactor: create mrinae-foundation package for separation of concerns
refactor: restructure styles and update imports

- Moved style files to a new package directory
- Updated import paths for colors, screens, and fonts
- Added mirinae-foundation as a dependency in package.json


refactor(tailwind): simplify config by extending default settings

- Replaced custom Tailwind configuration with an extended default config
- Removed redundant theme, variants, and plugins definitions
- Updated content paths for style processing
