+++ packages/mirinae-foundation/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "mirinae-foundation",
+  "private": true
+}
