+++ packages/postcss-config-custom/package.json
@@ -15,6 +15,7 @@
     "postcss-nested": "^4.2.3",
     "postcss-preset-env": "6.7.0",
     "postcss-simple-vars": "^5.0.2",
-    "tailwindcss": "^3.4.15"
+    "tailwindcss": "^3.4.15",
+    "mirinae-foundation": "*"
   }
 }
